bug fixes for session timeouts

+++ src/main/java/net/whydah/sso/session/WhydahAdminUserSession.java
@@ -45,7 +45,7 @@ public class WhydahAdminUserSession {
     }
 
     public static Integer calculateTokenRemainingLifetimeInSeconds(String userTokenXml) {
-        Integer tokenLifespanMs = UserTokenXpathHelper.getLifespan(userTokenXml);
+        Long tokenLifespanMs = UserTokenXpathHelper.getLifespan(userTokenXml);
         Long tokenTimestampMsSinceEpoch = UserTokenXpathHelper.getTimestamp(userTokenXml);
 
         if (tokenLifespanMs == null || tokenTimestampMsSinceEpoch == null) {
